Traktor: Implemented a "hash vector" in Model, used for vertices to improve "add unique" performance. Rearranged all cache directories, using a more homogenous structure. Optimized shader operations, no longer create a deep clone for several operations. Added "Flipboard" widget from GearUp into UiKit.
[git-p4: depot-paths = "//depot/traktor/": change = 24486]

+++ code/Model/Operations/SortCacheCoherency.cpp
@@ -18,7 +18,7 @@ bool SortCacheCoherency::apply(Model& model) const
 	AlignedVector< Polygon > newPolygons;
 	newPolygons.reserve(polygons.size());
 
-	uint32_t materialCount = model.getMaterials().size();
+	uint32_t materialCount = (uint32_t)model.getMaterials().size();
 	for (uint32_t material = 0; material < materialCount; ++material)
 	{
 		AlignedVector< uint32_t > indices;
